ansible: replace ssmtp with postfix and only using debian testing
There are tons of reasons why ssmtp should not be used [0],
so just get rid of it and replace it with postfix. It happens
to be much easier to read how we configure it as well.

The only reason we needed debian unstable was for ssmpt, since
we no longer need it, get rid of the debian package pinning stuff
for unstable.

We now rely purely on debian testing.

[0] https://www.davidgrant.ca/ssmtp_vs_postfix_no_contest

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>