roles: simplify oscheck's main task dependency chain
We can't run this right now:

  ansible-playbook -i hosts --tags run_tests oscheck.yml

This is because of the dependency chain requiring the install
to have happened successfully. The whole dependency chain thing
is just not needed given that tasks run linearly and stop if
one fails.

Just remove the superfluous dependency annotations as they
are not needed and just get in the way of using tags directly.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>