ansible: add support for parsing fstests bad files
We want to be able to very quickly be able to generate expunge
lists based on results directory where there are test cases which
have failed. To do this we look for all *.bad files, and generate
expunge files which oscheck can process.

To collect results:

  $ ansible-playbook -i hosts --tags copy_results oscheck.yml

Within the ansible directory after copying over the results you
can generate the expunge for a target new kernel with:

  $ ./python/generate_expunge_list.py xfs results/ ../expunges/

This also is used to update Debian's baseline for 4.19.0-4-amd64
using this mechanism. I had to first use any other 4.19.x baseline
so that at least the nodes would not crash.

Note that since I had used a 4.19.x baseline before for 4.19.0-4-amd64,
that means the next step I will have to concatenate the old baseline
with the new one and use uniq to remove duplicates.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>