ansible: print bad results as well in case xunit fails to be generated
In cases where the resulting xunit files fails to be produced
we're out of luck in trying to figure out what may have failed,
however the test case *.bad files can be used to infer which tests
failed.

We follow these premises:

  - We always wipe the results/ directory from the target guests we
    are testing
  - We always wipe our localhost results/ directory where we copy
    these results to
  - It is true that xunit can sometimes fail to be produced
  - Whenever a failure occurs this will be recorded on xunit files but
    only if there wasn't some sort of odd failure
  - Whenever a failure occurs this will be recorded on bad files, always

For this reason, bad files are our last go to for failures, which
can be used to expand our expunge files.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>