ansible/python: generate_expunge_list.py -> augment_expunge_list.py
Rename generate_expunge_list.py to augment_expunge_list.py and add
support so that we *only* add a new line into the file *iff* the
line we want to add does not exist.

This ensures that augment_expunge_list.py can be used *anytime* on
a run against our existing expunge list to augment it. The power of
this is that to be able to see if we ran into a regression we can
now just run:

	git diff

The regressions will appear as additions to the respective kernel's
expunge list. All a human would have to do after this is write a
bugzilla report and track it with a comment on the respective expunge
list.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>