expunges: add initial baseline for v4.19.58
We have no baseline for v4.19.58 yet, however we *know* we
cannot regress, as such nothing *new* from v4.19.18 should
be broken. We accept this assumption for an initial test using
a symlink. Let's see if this is true. If it is, we'll keep the
symlink to refect:

  * No regressions found

If the symlink is removed there is a delta observed. Note that
one of the things oscheck does not do yet:

  a) find tests which were expunged before but are now fixed
  b) find *new* tests coming in the fstests tree pipeline to
     ensure we're testing them. This can easily be fixed with a
     simple task.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>