README.md: update to reflect new features and TODO
A lot has changed. This is almost complete for doing
a full assessment of a filesystem with osheck's new
ansible / vagrant solution.

What's left?

Just to test it and provide fixes. The testing for v4.19.58
and the new proposed fixes *are* being tested with this framework
now. As such I suspect a future stable kernel is coming near you
soon with this whole ansible framework being used for it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>