oscheck: make default run all tests
I had made the default to run -g quick, but in practice I'm seeing
no results coming out. Although surprising, at this time, this is
producing running no tests at all. This could be that all quick tests
are expunged as they are known to fail, but that is far fetched to
believe.

In the meantime just diable -g quick by default for now.

While at it, I added oscheck.sh --fast-tests option  to let you
run the interpretation of what fast tests are as per oscheck.
Expand documentation to cover this and explain how one can use this
with ansible.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>