bootlinux: add support to manually uninstall a kernel version
Often, when doing Linux development, you may just want to nuke
a specific version of a kernel. Add support to do this. For
instance you may now run:

  ansible-playbook -i hosts --tags uninstall-linux --extra-vars "uninstall_kernel_ver=4.19.58" bootlinux.yml

This will remove 4.19.58, and update grub for you.
You may need to do this, for example, if after testing 4.19.58 you
realize you want to test an earlier version of the kernel. The
scripts which update grub tend to vary as to which kernel to use
as the default. Although grub has a support to let you specificy
the kernel you want, this mechanism is optional and removing a kernel
is sometimes just needed to save space on the drive as well.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>