expunges: update 4.19.18 baseline
After the kernel .config changes to include nvme support
and use it I noticed 85 regressions on 4.19.58 in comparison
to what we had on 4.19.18. I figured this could be a .config
change, and so I now tested the updated .config on 4.19.18
and these are the results that come in.

Doing a diff between 4.19.18 and 4.19.58 and only looking
at the additions tells us the new candidate regressions.

I then the moved the target kernel back onto 4.19.18 and tested
with naggy-check.sh each candidate regression. So far it seems
there is only one real possible regression. Details below.

generic/418 was passing on 4.19.18, but after a failure was picked
up on 4.19.58 I ran naggy-check.sh with it on 4.19.18. It failed
after 10 runs in a row. So this is not a regression.

The next candidate regression was generic/558, but this is just
a rename of shared/006 through commit a9868eb48d62 ("shared,generic:
move shared/006 to generic/"). But shared/006 was only deteted failing
on:

```
$ git grep "shared/006"
expunges/any/large-disk.txt:shared/006
expunges/any/over-10s.txt:shared/006 # takes 70s
expunges/linux-next/xfs/unassigned/all.txt:shared/006
```

So we know it runs for a while, and we had that on linux-next
it failed as well. That's all.

But in the testing done it was detected to fail on these section:

  * xfs_nocrc
  * xfs_nocrc_512
  * xfs_reflink

I've run naggy-check.sh on 4.19.18 with generic/558 on these sections,
and so far I see no failures. The naggy-check.sh infinite loop
test (naggy-check.sh -s $section -f generic/558) is still running
on each respective host, however these are the number so far:

generic/558
        xfs_nocrc_512 - PASS 316
        xfs_nocrc - PASS 363
        xfs_reflink - PASS 1130

So confidence seems pretty high at this point that if generic/558 really
is failing on 4.19.58 it is a regression.

To make regression testing easier we need to have double amount of
hosts. One set for development and one for stable. This will allow us
to easily go back and forward without having to switch kernels
to confirm a regression. That effort will be done next.

Note: yes the result of the diffs between 4.19.18 and 4.19.58 also
reveal that 4.19.58 may have some issues fixed. However for this
confirmation is best to instead create a new specialized task which goes
through *each* test on an expunge list, and naggy-checks.sh's it to
confirm the issue is still present. That is, we need a a 'fix finder'
task. But this will be a be a bit complex as it means we'll need a
ping task, which if the ping fails we'll reboot. Or.. we need to verify
the watchdog actually works.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>