vagrant/ansible: add development hosts
Testing fstests with oscheck requires being able to *confirm*
regressions. To be able to do this easily we actually need one
set of hosts running the stable kernels, and another set of hosts
running the development kernel we are testing.

So augment the hosts we create with a group of development hosts.
Just add a "-dev" postfix to them. Fortunately ansible lets us group
up hosts together easily, and lets us run playbooks against only a
group with -l. For example:

  ansible-playbook -i hosts -l dev oscheck.yml

This means the section on the task which runs oscheck.sh must also trim
this postfix.

With all this in place we now can easily run tests against stable and
development kernels concurrently. To change the target kernel or repo,
on can just use --extra-vars, for instance:

 ansible-playbook -i hosts -l baseline --extra-vars "target_linux_version=4.19.18" bootlinux.yml

That will deploy 4.19.18 onto the baseline hosts.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>