expunges: update generic/388 failures
generic/388 will trigger a crash on xfs_nocrc, xfs_reflink with the
same trace. On the latest upstream code we get a different crash with
reflink if we disable rmapbt and use reflink.

We're currently not tracking such type of configuration, but provided
Zorro justifies why he's testing with it, we may want to add a new
configuration to track that setup. A "xfs_reflink_normapbt"
configuraiton, if you will.

The following bugs are tracking these issues:

https://bugzilla.kernel.org/show_bug.cgi?id=204223 - 4.19.58
https://bugzilla.kernel.org/show_bug.cgi?id=204049 - xfs-linux xfs-5.3-merge-6

The kz# annotations on expunges are shorthands for bugzilla.kernel.org
entries.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>