kdevops/vagrant: use libvirt-qemu group as default and allow override
Debian uses libvirt-qemu, so use that as the default given all defaults
are set for Debian. Fedora uses the qemu group, you can override with
the OSCHEK_VAGRANT_QEMU_GROUP=qemu environment variable.

This is a port of kdevops commit a10fa72425c ("vagrant: use libvirt-qemu
group as default and allow to override").

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>