expunges: add 4.19.62 baseline to match v4.19.58 for now
We have a few suspect fixes, such as the kernel crash which is triggered
on generic/388, reported via kz#204223 [0] and generic/530 which should
no longer fail (pending testing and confirmation of course). However,
its best to keep the baseline and only remove entries once we have
full empirical test cases which clearly annotate *why* a test was fixed
and we document how many times we've ran a test for.

The fact that the baseline of v4.19.62 is identical to v4.19.58 is
reflected in code as this is a symlink right now. Once and if we
diverge this can change.

[0] https://bugzilla.kernel.org/show_bug.cgi?id=204223

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>