terraform: update ansible role to get ssh config updates
The terraform ansible role has been updated to add support for
automatically updating your ~/.ssh/config with the target hosts
for a slew of cloud providers. Bump the release to allow us to
take advtange of that new feature.

Tested with AWS with the latest terraform as of today. That is:

terraform v0.12.19

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>