oscheck_nodes.yaml: adjust as per new role updates
The ssh key path is pointless as we always generate a new random one,
and update your ssh config for you now. Using a static ssh key can be
a security risk as well, if someone steals that *one* key they have
access to all your systems you have created via vagrant. A randomized
key makes more sense for random ephemeral systems.

We've also put quite a bit of effort into ensuring we automate your
libvirt setup so you don't have to do things, and hence the new
qemu_group_auto.

We remove the setting of machine_type to q35 as q35 can mean
different things depending on *which* release of qemu you use
and so just using the sensible defaults there should suffice for
the era you need to use. If you use q35 that may not work for
older releases. The rationale behind this is explained on the
kdevops_vagrant commit 003ad8cc9756d [0], which we'll quote here
so that it is is clear:

	Allow the qemu machine type to be box specific and be skipped

	Using the default of q35 made some sense when I was first testing things
	out with the latest linux-next releases, it however doesn't make much
	sense when considering really old releases.

	These days, as of qemu-4.2.0 the machine type value of
	q35 is an alias pc-q35-4.2. However for older qemu releases, say circa
	SLE11-SP4 days q35 was an alias pc-q35-1.4. If you are strictly using
	the machine type, you'll then need to also be able to specify a
	different one than the global one for older systems.

	For instance SLE12-SP1 cannot boot pc-q35-4.2, at least its first
	kernel cannot. So you'd need to set it to pc-i440fx-4.2. Doing
	this is tedious, so if you run into these situations just unset
	the machine type.

	You should use virt-install --os-variant specified so you get the right
	settings. Get the variant with $(osinfo-detect os), or osinfo-detect
	against the ISO file. That will set the right settings for your guest.

	vagrant relies on booting a qcow2 image without any of the information
	used during virt-install, however somehow the right value is used. So
	if you do not have a strong reason to force the machine type, don't set
	it.

	Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>

Other than this we have learned also that you can specify sensible
defaults for qemu on the vagrant box you use, and so the onus falls
on the vagrant box creator to ensure defaults that work are put in
place there.

Lastly, we remove the dhcp hack as the issue is no longer present on
the latest vagrant debian box, which originally had motivated the
addition of that.

[0] https://github.com/mcgrof/kdevops_vagrant/commit/003ad8cc9756db6e15dd34e6e06ce9e973c9b543

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>