requirements: bump to kdevops 1.9.8
This update kdevops to v1.9.8 so that we can take
advatange of specifyng the ansible python interpreter.
Without this we see messages complaining about apt-python
package missing. The reason this happens is due to the
older default interpreter detected and preferred. While
this makes it easier to use on older systems, it puts the
burden then on using the new interpreter on newer systems.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>